add appbar component, need to perfect logout
