--- client/src/components/Jumbotron/index.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-function Jumbotron({ children }) {
-  return (
-    <div
-      style={{ height: 560, clear: "both", paddingTop: 120, textAlign: "center" }}
-    >
-      {children}
-    </div>
-  );
-}
-
-export default Jumbotron;
