+++ client/src/pages/Profile.js
@@ -1,0 +1,19 @@
+import * as React from 'react';
+
+import Box from '@mui/material/Box';
+// import Typography from '@mui/material/Typography';
+
+
+const Profile = () => {
+
+  return (
+    <Box>
+
+    </Box>
+  );
+};
+
+export default Profile;
+
+
+
