+++ client/src/utils/auth.js
@@ -31,7 +31,7 @@ class AuthService {
     // Saves user token to localStorage
     localStorage.setItem('id_token', idToken);
 
-    window.location.assign('/');
+    window.location.assign(`/profile`);
   }
 
   logout() {
